MODFISTO-90 Support all-or-nothing operations for pending payments by invoice
- extract common transactions DAO logic to BaseTransactionsDAO
